% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myApp.R
\name{myApp}
\alias{myApp}
\title{sprtt shiny app}
\usage{
myApp(...)
}
\arguments{
\item{...}{XXX}
}
\value{
Opens the Web App
}
\description{
Shiny weg app of the sprtt package to calculate sequential
t-tests.
}
